+++ CHANGELOG.md
@@ -23,6 +23,8 @@ nav_order: 1
 - Replaced `aiven-go-client/v2` with `aiven/go-client-codegen` in `account_team` resource/data source
 - Replaced `aiven-go-client/v2` with `aiven/go-client-codegen` in `account_team_member` resource/data source
 - Replaced `aiven-go-client/v2` with `aiven/go-client-codegen` in `account_team_project` resource/data source
+- Replaced `aiven-go-client/v2` with `aiven/go-client-codegen` in `aiven_pg_user` resource/data source
+- Fix `aiven_pg_user` creating in bulk occasionally results in a 404 error
 
 ## [4.31.1] - 2024-12-23